*Solving Blocking Code
Design Context:
The User Interface was designed to await for the user to choose to submit all the products. Once pressed, the final report would then be drawn in a different JFrame. However, because I also designed it to await the Report to close before triggering the main Interface's GUI's latch, it would stay blocked. Maybe swing's executor loses authority to execute the Report Frame, and therefore it coudln't continue drawing the report due to the awaits.

The solution I came up was to expose the latch of the report window to the main UI object, so I could await the inner latch in the mainthread.
What's my rating of the technical debt incurred in this?
It's in the level of, you might have to rewrite everything, because I was pushed to a corner and had to write something that I assumed could solve the blocking of the code.

Fortunately, the interfaces I had written for the "Inventory" should be reusable without issues.